+++ x/oracle/types/genesis.go
@@ -26,16 +26,16 @@ func ValidateGenesis(data GenesisState) error {
 		return fmt.Errorf("relayer timeout should be positive, is %d", data.Params.RelayerTimeout)
 	}
 
-	if data.Params.RelayerBackoffTime <= 0 {
-		return fmt.Errorf("the relayer backoff time must be positive, is %d", data.Params.RelayerBackoffTime)
-	}
-
 	if data.Params.RelayerRewardShare <= 0 {
 		return fmt.Errorf("the relayer reward share must be positive, is %d", data.Params.RelayerRewardShare)
 	}
 
 	if data.Params.RelayerRewardShare > 100 {
 		return fmt.Errorf("the relayer reward share should not be larger than 100, is %d", data.Params.RelayerRewardShare)
+	}
+
+	if data.Params.RelayerInterval <= 0 {
+		return fmt.Errorf("the relayer interval should be positive, is %d", data.Params.RelayerInterval)
 	}
 	return nil
 }